+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # UNRELEASED
 
  - menu bar now allows for navigation with arrow keys
+ - introduced a table container element 
 
 # 0.1.0